refactor(context): extract bundle manifest writer
